[Day 04] Linear parallelization is still parallelization

+++ day00/kernel.cl
@@ -1,13 +1,13 @@
 __kernel void part1(
-    __global const long* inp,
-    __global long* out
+    __global const int* inp,
+    __global int* out
 ) {
     out[0] = 1337;
 }
 
 __kernel void part2(
-    __global const long* inp,
-    __global long* out
+    __global const int* inp,
+    __global int* out
 ) {
     out[0] = 1338;
 }
